Make OAuth request if credentials are available
Some methods (like user shelves) use app authentication (requests that
send token=SECRET_TOKEN). This works if a user is public, but if a user
has restricted who can view their information you need to use
`Goodreads::Request#oauth_request`. Rather than have the caller specify
which to use, we can infer the desired behavior based on whether the
client was instantiated with an API key or an OAuth token.